fix reviewing issues
