+++ .umirc.ts
@@ -1,46 +1,13 @@
 import { defineConfig } from 'dumi';
+import { getLocalStorageProxy } from '@nocobase/utils';
 
 process.env.MFSU_AD = 'none';
 
 const {
   API_ORIGIN = '',
   API_ORIGIN_DEV,
-  API_BASE_PATH = '/api/',
-  LOCAL_STORAGE_USE_STATIC_SERVER,
-  LOCAL_STORAGE_BASE_URL
+  API_BASE_PATH = '/api/'
 } = process.env;
-
-function getLocalStorageProxy() {
-  if (!LOCAL_STORAGE_USE_STATIC_SERVER) {
-    return {};
-  }
-
-  const baseUrl = LOCAL_STORAGE_BASE_URL || '/uploads';
-  let url;
-  try {
-    url = new URL(baseUrl);
-  } catch(e) {
-    url = {
-      href: `${API_ORIGIN_DEV}${baseUrl}`
-    };
-  }
-
-  // if local storage url is some network address,
-  // such as ip or domain name rather than API_ORIGIN,
-  // do not proxy.
-  if (!API_ORIGIN_DEV || !url.href.startsWith(API_ORIGIN_DEV)) {
-    return {};
-  }
-
-  const config = {
-    [baseUrl]: {
-      target: API_ORIGIN_DEV,
-      changeOrigin: true,
-    }
-  }
-
-  return config;
-}
 
 export default defineConfig({
   title: 'NocoBase',
