+++ packages/create-nocobase-app/templates/AppGenerator/.umirc.ts
@@ -1,6 +1,7 @@
 import { defineConfig } from 'umi';
 import dotenv from 'dotenv';
 import path from 'path';
+import { getLocalStorageProxy } from '@nocobase/utils';
 
 dotenv.config({
   path: path.resolve(__dirname, './.env'),
@@ -11,42 +12,8 @@ process.env.MFSU_AD = 'none';
 const {
   API_ORIGIN = '',
   API_ORIGIN_DEV,
-  API_BASE_PATH = '/api/',
-  LOCAL_STORAGE_USE_STATIC_SERVER,
-  LOCAL_STORAGE_BASE_URL
+  API_BASE_PATH = '/api/'
 } = process.env;
-
-function getLocalStorageProxy() {
-  if (!LOCAL_STORAGE_USE_STATIC_SERVER) {
-    return {};
-  }
-
-  const baseUrl = LOCAL_STORAGE_BASE_URL || '/uploads';
-  let url;
-  try {
-    url = new URL(baseUrl);
-  } catch(e) {
-    url = {
-      href: `${API_ORIGIN_DEV}${baseUrl}`
-    };
-  }
-
-  // if local storage url is some network address,
-  // such as ip or domain name rather than API_ORIGIN,
-  // do not proxy.
-  if (!API_ORIGIN_DEV || !url.href.startsWith(API_ORIGIN_DEV)) {
-    return {};
-  }
-
-  const config = {
-    [baseUrl]: {
-      target: API_ORIGIN_DEV,
-      changeOrigin: true,
-    }
-  }
-
-  return config;
-}
 
 export default defineConfig({
   favicon: '/favicon.png',
