+++ packages/plugin-file-manager/package.json
@@ -4,7 +4,6 @@
   "main": "lib/index.js",
   "license": "MIT",
   "dependencies": {
-    "@koa/cors": "^3.1.0",
     "@koa/multer": "^3.0.0",
     "@nocobase/server": "^0.5.0-alpha.34",
     "ali-oss": "^6.12.0",
