+++ packages/utils/src/index.ts
@@ -1,9 +1,2 @@
-import deepmerge from 'deepmerge';
-
-const overwriteMerge = (destinationArray, sourceArray, options) => sourceArray
-
-export function merge(obj1: any, obj2: any) {
-  return deepmerge(obj1, obj2, {
-    arrayMerge: overwriteMerge,
-  });
-}
+export * from './merge';
+export * from './getLocalStorageProxy';
