+++ packages/utils/src/merge.ts
@@ -1,0 +1,9 @@
+import deepmerge from 'deepmerge';
+
+const overwriteMerge = (destinationArray, sourceArray, options) => sourceArray
+
+export function merge(obj1: any, obj2: any) {
+  return deepmerge(obj1, obj2, {
+    arrayMerge: overwriteMerge,
+  });
+}
